chatbot-requirements.txt
nltk
newspaper3k
